requirements.txt
appnope==0.1.2
APScheduler==3.6.3
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.9.3
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.4.6
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
et-xmlfile==1.0.1
Flask==1.1.2
idna==2.8
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.19.5
openpyxl==3.0.7
packaging==20.9
pandas==1.2.0
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.18
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
PyJWT==1.7.1
pymongo==3.11.3
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-telegram-bot==13.2
pytz==2020.5
pyupbit==0.2.8
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
reprint==0.6.0
requests==2.21.0
scikit-learn==0.24.1
scipy==1.6.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
soupsieve==2.2
tabulate==0.8.9
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
tzlocal==2.1
upbitpy==1.0.1
urllib3==1.24.3
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1